% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.vectors_cache_path}
\alias{.vectors_cache_path}
\title{Return the default cache path for the dining vectors RDS}
\usage{
.vectors_cache_path(cache_dir = rappdirs::user_data_dir("dineExpand"))
}
\description{
Return the default cache path for the dining vectors RDS
}
\keyword{internal}
